Progress on app layout.
