+++ .vscode/settings.json
@@ -35,4 +35,8 @@
   "[html]": {
     "editor.defaultFormatter": "esbenp.prettier-vscode",
   },
+
+  "[css]": {
+    "editor.defaultFormatter": "esbenp.prettier-vscode"
+  },
 }
