+++ apps/gui/src-tauri/src/main.rs
@@ -1,15 +1,19 @@
 // Prevents additional console window on Windows in release, DO NOT REMOVE!!
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
-// Learn more about Tauri commands at https://tauri.app/v1/guides/features/command
 #[tauri::command]
-fn greet(name: &str) -> String {
-  format!("Hello, {}! You've been greeted from Rust! Pissballs!", name)
+fn start_extraction() {
+  println!("start_extraction() running");
+}
+
+#[tauri::command]
+fn do_refresh() {
+  println!("do_refresh() running");
 }
 
 fn main() {
   tauri::Builder::default()
-    .invoke_handler(tauri::generate_handler![greet])
+    .invoke_handler(tauri::generate_handler![start_extraction, do_refresh])
     .run(tauri::generate_context!())
     .expect("error while running tauri application");
 }
