+++ apps/gui/src/styles.css
@@ -13,4 +13,3 @@
   -moz-osx-font-smoothing: grayscale;
   -webkit-text-size-adjust: 100%;
 }
-
